+++ .gitignore
@@ -113,7 +113,7 @@ ENV/
 env.bak/
 venv.bak/
 config.ini
-tpe_daytrip.sql
+db_tpetrip.sql
 
 # Spyder project settings
 .spyderproject
